Renaming MainActivity. Extracting FragmentManager.
