+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
         android:label="@string/app_name"
         android:theme="@style/Theme.DesignDemo">
 
-        <activity android:name=".MainActivity">
+        <activity android:name=".DrawerActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.DEFAULT" />
@@ -33,9 +33,9 @@
         </activity>
 
         <activity android:name=".CheeseDetailActivity"
-            android:parentActivityName=".MainActivity">
+                  android:parentActivityName=".DrawerActivity">
             <meta-data android:name="android.support.PARENT_ACTIVITY"
-                android:value=".MainActivity" />
+                       android:value=".DrawerActivity"/>
         </activity>
 
     </application>
